Refactor code to simplify and make more usable else where
